Make ambiguous columns in order by expressions be errors.
In HSQLDB there is some code to replace aliases in the order by list
with the expressions which the aliases name.  This code had the same
diseases the regular column reference resolution code did, in that it
stopped when it found one definition.  This means that ambiguous
definitions are not discovered.

This fix finds ambiguous column references in the order by list and
throws errors.  Tests were added for group by, order by and having.
There were already several tests for where and joins.

There are still some problems with aliases in group by expressions.
These are to be addressed in ENG-9782.

https://issues.voltdb.com/browse/ENG-9828